+++ example/data_source_projects.tf
@@ -3,12 +3,12 @@ data "ironio_projects" "example" {
   filter_name = "${ironio_project.example.name}"
 }
 
-output "ironio_projects_example_ids" {
+output "data_ironio_projects_example_ids" {
   description = "The project ids"
   value       = "${data.ironio_projects.example.ids}"
 }
 
-output "ironio_projects_example_names" {
+output "data_ironio_projects_example_names" {
   description = "The project names"
   value       = "${data.ironio_projects.example.names}"
 }