+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 BREAKING CHANGES:
 
 * provider: Changed configuration schema to use blocks (`auth`, `cache`, `mq` and `worker`)
-* provider: Changed the default value of `load_config_file` to `true`
+
+FEATURES:
+
+* **New Data Source:** `ironio_pull_queue`
+* **New Data Source:** `ironio_push_queue`
 
 ENHANCEMENTS:
 
 * makefile: Added `example` target to simplify the testing process
+* resource/ironio_pull_queue: Added `message_count` and `message_count_total` attributes
+* resource/ironio_push_queue: Added `message_count` and `message_count_total` attributes
 
 ## 0.1.0